regroup things related to public-key cryptography
