+++ cmd/storj-sim/inmemory.go
@@ -19,7 +19,7 @@ import (
 
 	"storj.io/storj/internal/processgroup"
 	"storj.io/storj/internal/testplanet"
-	"storj.io/storj/pkg/peertls"
+	"storj.io/storj/pkg/pkcrypto"
 )
 
 func inmemoryRun(flags *Flags) error {
@@ -101,14 +101,14 @@ func inmemoryTest(flags *Flags, command string, args []string) error {
 		}
 
 		var chainPEM bytes.Buffer
-		errLeaf := pem.Encode(&chainPEM, peertls.NewCertBlock(identity.Leaf.Raw))
-		errCA := pem.Encode(&chainPEM, peertls.NewCertBlock(identity.CA.Raw))
+		errLeaf := pem.Encode(&chainPEM, pkcrypto.NewCertBlock(identity.Leaf.Raw))
+		errCA := pem.Encode(&chainPEM, pkcrypto.NewCertBlock(identity.CA.Raw))
 		if errLeaf != nil || errCA != nil {
 			return errs.Combine(errLeaf, errCA, planet.Shutdown())
 		}
 
 		var key bytes.Buffer
-		errKey := peertls.WriteKey(&key, identity.Key)
+		errKey := pkcrypto.WriteKey(&key, identity.Key)
 		if errKey != nil {
 			return errs.Combine(errKey, planet.Shutdown())
 		}
